Renamed working branch to dev/sprint, finish sprint 4 (#310)
* Fix for 'Show Solutions'. Correct answers are now displayed, and checkmarks and crosses only shown for logged in answers.

* Fix for 'Show Solutions'. Correct format is now displayed.

* Fixed last updated, preview and simple text display and simplified function diffSolution.

* If no answer is given, correct answers will be marked with red cross as incorrect.

* Readability improvements.

* Changed correct checkbox to radio buttons, should now also display solutions. No conditional between SC and MC since unnecessary.

* TLI-71 corrected own bad advise

* TLI-1 correct answer is crossed when no answer is given

* TLI-1 solution array now uses boolean

* Finish sprint 4, updated version, CHANGELOG

* mirrored view of edit.vue in create.vue, added mixin for shared methods
