+++ CHANGELOG
@@ -1,5 +1,9 @@
 # Changes
 
+## v1.2.4
+- fixed wrong display of solutions in assessments
+- changed display of choice of correct solution in SCMC component
+
 ## v1.2
 - moved http calls to vuex
 - added more unit tests
