+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "laya-iraq-frontend",
-  "version": "1.2.0",
+  "version": "1.2.4",
   "description": "Frontend For LAYA",
   "private": true,
   "scripts": {
