+++ src/plugins/laya-la-drag-drop/view.vue
@@ -127,7 +127,10 @@ Dependencies:
         v-for="(item, index) in items"
         :key="index"
       >
-        {{ item.label }}: {{ categories[items[index].category] }},
+        {{ item.label }}: {{ courseSimple
+          ? categories[item.category].simple
+          : categories[item.category].text
+        }}
       </div>
     </div>
   </div>
