+++ src/plugins/laya-la-scmc/create.vue
@@ -235,13 +235,14 @@ Dependencies: @/mixins/locale.vue
           </div>
 
           <!-- correct -->
-          <div class="form-check form-check-inline">
+          <div class="form-check form-check-inline ">
             <input
               :id="'option-corr-'+i"
               v-model="solutions[i]"
               class="form-check-input"
-              type="checkbox"
-              :true-value="i"
+              :type="multiple ? 'checkbox' : 'radio'"
+              :value="true"
+              @click="() => { if (!multiple) {for (const ix in solutions) { ix === i ? solutions[ix] = true : solutions[ix] = false } } }"
             >
             <label
               class="form-check-label"
@@ -300,11 +301,13 @@ Dependencies: @/mixins/locale.vue
 <script>
 import { locale, tooltipIcon } from '@/mixins'
 import { v4 as uuidv4 } from 'uuid'
+import commonMethods from './choices-common-methods'
 
 export default {
   name: 'LayaLaScmcCreate',
 
   mixins: [
+    commonMethods,
     locale,
     tooltipIcon
   ],
@@ -349,32 +352,6 @@ export default {
         flagged: false,
         id: uuidv4()
       }
-    },
-
-    /**
-     * Function _delItem(idx): Remove an item at given index
-     *
-     * Author: core
-     *
-     * Last Updated: unknown
-     *
-     * @param {*} idx index at which the item is to be removed
-     */
-    _delItem (idx) {
-      this.options.splice(idx, 1)
-    },
-
-    /**
-     * Function _addItem: add a new item to options
-     * Author: core
-     * Last Update: June 28, 2021
-     */
-    _addItem () {
-      this.options.push({
-        text: this.y18n('layaLaScmc.edit.sampleOption'),
-        flagged: false,
-        id: uuidv4()
-      })
     }
   }
 }
